Fix bazel command for BLAKE3 asm on windows
This adds a new condition to the existing select() found in the target `//llvm:Support`. This changes to use the windows assembly source files when compiling on windows.

Note that this is not a perfect fix, as there exists assembly source files for both msvc and gnu on windows, and this just assumes msvc is being used when using bazel to compile on windows. Unfortunately the `@bazel_tools//src/conditions:windows_msvc` config_setting does not appear to work, otherwise that would be used instead. While this is not a perfect solution, this should handle the most likely use case, fixing the fact that compiling on windows was simply broken before. Existing unix support is unchanged.